prompt user to confirm room deletion

+++ desktop-frontend/src/components/RoomModal.tsx
@@ -226,6 +226,14 @@ const RoomModal = ({
           <button
             className=" mr-auto ml-auto p-2 border-r-black border-l-black border-t-black border-b-black rounded-xl border-2 shadow-lg inline-block"
             onClick={async () => {
+              let isConfirmed = confirm(
+                "Are you sure you want to delete this room?"
+              );
+
+              if (!isConfirmed) {
+                return;
+              }
+
               await deleteRoom(roomData.id);
             }}
           >
